refactor: move checkMinimumReleaseAge to lib/util/minimum-release-age.ts
